1_crewAi/requirements.txt
crewai>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
litellm>=1.79.3
